prototype works on upsquaredpro, x64 win10, bug fix
Bug fix: increase rsi dsi adress by 16, to be accurate with
the orignal cmps instruction. Now it works for both x86 and
x64 Win10.

The cpu Intel N4200 on upsquaredpro is newer than the one on
upsquared. There are space in msram and match/patch.

The in fix point was at 0x58ba.

    hook_match_and_patch(0x1f, 0x58ba, 0x017a);

Use slot 30 for cmps hook.

Microcode is placed staring at 0x7d30.